+++ web/api/metadata_importer_multipart.php
@@ -1,0 +1,10 @@
+<?php
+/*
+ Rqwatch
+ Copyright (C) 2025 Giannis Kapetanakis
+
+ This Source Code Form is subject to the terms of the Mozilla Public
+ License, v. 2.0. If a copy of the MPL was not distributed with this
+ file, You can obtain one at http://mozilla.org/MPL/2.0/.
+*/
+require __DIR__ . '/../../app/Api/metadata_importer_multipart.php';